Refactor FollowFeed
